refactor: extract EvolutionModalHelper to resolve max file and lines eslint constraints
